No fucking clue what I'm doing here, definitely not working
