+++ .gitignore
@@ -1,3 +1,2 @@
 .env
-package-lock.json
 node_modules
