+++ package.json
@@ -1,10 +1,12 @@
 {
-  "name": "CSBotSlash",
+  "name": "csbotslash",
   "version": "1.0.0",
   "description": "",
-  "main": "src/index.js",
+  "main": "dist/index.js",
+  "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "npx tsc -p ./ && npm run start",
+    "start": "node ."
   },
   "keywords": [],
   "author": "",
@@ -13,6 +15,16 @@
     "discord.js": "github:vaporox/discord.js#feat-interactions",
     "dotenv": "^8.2.0",
     "get-cat-facts": "0.0.3",
-    "random-stuff-api": "^10.0.0"
+    "random-stuff-api": "^10.0.0",
+    "tslib": "^1.14.1"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^4.22.0",
+    "@typescript-eslint/parser": "^4.22.0",
+    "eslint": "^7.24.0",
+    "eslint-config-standard": "^16.0.2",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.3.1"
   }
 }
