+++ src/@types/get-cat-facts.d.ts
@@ -1,0 +1,7 @@
+declare module 'get-cat-facts' {
+    export const random: () => Item[];
+
+    interface Item {
+        text: string;
+    }
+}
