--- src/commands/ping.js
@@ -1,12 +1,0 @@
-module.exports = {
-    name: 'ping',
-    create: {
-        data: {
-            name: 'ping',
-            description: 'Ping Pong!'
-        }
-    },
-    execute(client, interaction) {
-        interaction.editReply(':tada: | Still alive!');
-    }
-}
